feat(cli): pull latest agent version from S3
