+++ cli/cmd/cli_state_test.go
@@ -1,0 +1,34 @@
+//
+// Author:: Salim Afiune Maya (<afiune@lacework.net>)
+// Copyright:: Copyright 2020, Lacework Inc.
+// License:: Apache License, Version 2.0
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+package cmd
+
+import (
+	"github.com/lacework/go-sdk/lwlogger"
+)
+
+func init() {
+	if cli.Log == nil {
+		// initialize the Lacework logger for testing purposes
+		//
+		// this init() function is needed only because the initConfig() func
+		// never runs when we are testing regular functions, it is only called
+		// when cobra runs a CLI command
+		cli.Log = lwlogger.New(cli.LogLevel).Sugar()
+	}
+}
